requirements.txt
pyyaml==6.0.2
openai>=1.53.0
dashscope>=1.17.0
jsonschema>=4.22.0
pytest>=8.0.0
